fix of a missing state, removal of unnecessary test files, fix in config wizard
most removed tests are not relevant for germany as they was related in
the clone from netherlands to invoice basis (related to invoice or
account move dates which is probably not relevant for germany). The
change in the config wizard was required as the tax tag mappings for
the tax statement should work with skr03 and skr04 accounting plans.