adoption to UStVA Report 2019
The old lines 39 disappeared and was merged together with the old line
40 (now they are line 39 in 2019 edition). Furthermore the old lines 49
and 51 disappeared and was merged into the line 50. There is a
relevance for the tax codes (tag_ids). The old tag „USt-VA KZ 68“ have
to be mapped to the same line as the tag „USt-VA KZ 60“. The old tags
„USt-VA KZ 52“ and „USt-VA KZ 78“ have to be mapped the same way as the
old and single remaining „USt-VA 84“ (= base values). The old tags
„USt-VA KZ 53“ and „USt-VA KZ 79“ have to be mapped the same way as the
old and single remaining „USt-VA KZ 85“ (tax values). Furthermore we
might have formatting changes caused by the different line numbering.
Therefore i expect further adoptions in the files
„l10n_de_tax_statement_line.py“ and also in the
„l10n_de_tax_statement_config_wizard.xml“ and the
„l10n_de_tax_statement_config_wizard.py“. I assume we cannot avoid
deeper changes in order to adopt what we really want.